Updated to reflect the matrix skew feature in CAVE

+++ mechanica/src/main/kotlin/com/mechanica/engine/drawer/subclasses/transformation/TransformationDrawerImpl.kt
@@ -17,7 +17,7 @@ class TransformationDrawerImpl(drawer: Drawer, private val state: DrawState): Tr
     }
 
     override fun skew(horizontal: Radian, vertical: Radian): TransformationDrawer {
-        state.setSkew(horizontal.toFloat(), vertical.toFloat())
+        state.setSkew(horizontal, vertical)
         return this
     }
